fix simple I/O on large files

+++ src/main/java/edu/berkeley/bid/UTILS.java
@@ -10,13 +10,13 @@ public final class UTILS {
         System.loadLibrary("bidmatmkl");
     }
 
-    public static native  void memcpybi( int n, byte [] a, int startA, int [] b, int startB );
-    public static native  void memcpybf( int n, byte [] a, int startA, float [] b, int startB );
-    public static native  void memcpybd( int n, byte [] a, int startA, double [] b, int startB );
+    public static native  void memcpybi( int n, byte [] a, long startA, int [] b, long startB );
+    public static native  void memcpybf( int n, byte [] a, long startA, float [] b, long startB );
+    public static native  void memcpybd( int n, byte [] a, long startA, double [] b, long startB );
 
-    public static native  void memcpyib( int n, int [] a, int startA, byte [] b, int startB );
-    public static native  void memcpyfb( int n, float [] a, int startA, byte [] b, int startB );
-    public static native  void memcpydb( int n, double [] a, int startA, byte [] b, int startB );
+    public static native  void memcpyib( int n, int [] a, long startA, byte [] b, long startB );
+    public static native  void memcpyfb( int n, float [] a, long startA, byte [] b, long startB );
+    public static native  void memcpydb( int n, double [] a, long startA, byte [] b, long startB );
     
     public static OutputStream _getOutputStream(String fname, Boolean compressed, int compressionLevel) throws IOException {
     	FileOutputStream fout = new FileOutputStream(fname);
